stress-tests/python-scripts/requirements-simple.txt
# Dependencias simplificadas para pruebas de estrés
locust==2.17.0
selenium==4.15.0
webdriver-manager==4.0.1
faker==20.1.0
requests==2.31.0
firebase-admin==6.2.0
google-cloud-firestore==2.13.1
psutil==5.9.6
python-dotenv==1.0.0
